+++ clmm/__init__.py
@@ -1,8 +1,9 @@
 """ CLMM is a cluster mass modeling code. """
 from .gcdata import GCData
 from .galaxycluster import GalaxyCluster
-from .polaraveraging import compute_tangential_and_cross_components, make_binned_profile
-from .utils import compute_radial_averages, make_bins, convert_units
+from .hybrid import make_binned_profile, convert_units
+from .polaraveraging import compute_tangential_and_cross_components
+from .utils import compute_radial_averages, make_bins
 from .modeling import cclify_astropy_cosmo, get_reduced_shear_from_convergence, get_3d_density, predict_surface_density, predict_excess_surface_density, angular_diameter_dist_a1a2, get_critical_surface_density, predict_tangential_shear, predict_convergence, predict_reduced_tangential_shear
 
 from . import lsst
